command: alias video/audio/sub properties to -vid/-aid/-sid
Now vid/aid/sid can be used as properties. video/audio/sub still work,
but they are aliases for the "real" properties.

This guarantees that options/properties use the same value range. Once
consequence is that the video/audio/sub properties return "no" as value
if no track is selected instead of -1.